Personal website practice restart
I had to restart this project. I followed Google instructions to figure out why the original project wouldn't upload to Github, and it replaced my files with the ones in my "coding-practice" repository. I edited resource code from a website tutorial I was learning from to get as close to as I could to where I was, and that is where we are now. The rest past this commit will be personally typed code, and I got to this point by myself previously. By the end, this will be a fully-functioning site with information relevant to myself.